refactor: Move dlc handler to dlc module
