+++ mobile/native/src/ln_dlc/mod.rs
@@ -8,8 +8,8 @@ use crate::backup::DBBackupSubscriber;
 use crate::commons::reqwest_client;
 use crate::config;
 use crate::db;
-use crate::dlc_handler;
-use crate::dlc_handler::DlcHandler;
+use crate::dlc::dlc_handler;
+use crate::dlc::dlc_handler::DlcHandler;
 use crate::event;
 use crate::event::EventInternal;
 use crate::ln_dlc::node::Node;
